requirements.txt
# Project dependencies for College AI multithreaded crawler with Zilliz Cloud integration
# Core web scraping and data processing
requests>=2.31.0  # HTTP library for web requests
beautifulsoup4>=4.12.2  # HTML/XML parser for web scraping
lxml>=4.9.4  # fast HTML parser for BeautifulSoup (recommended)
pyyaml>=6.0.1  # for per-domain Playwright profiles

# Database and vector operations
pymilvus>=2.5.0  # for Zilliz Cloud / Milvus with hybrid search (dense + BM25)
sqlalchemy-libsql>=0.1.0  # Turso/libSQL dialect for SQLAlchemy

# AI and embeddings
openai>=1.14.2  # for OpenAI embeddings and API access
tiktoken>=0.7.0  # token-based chunking for RAG
cohere>=5.0.0  # cross-encoder reranking (Cohere rerank-v4.0-pro)
rapidfuzz>=3.0.0  # fuzzy school name matching in RAG router

# Browser automation (JS rendering and bot detection bypass)
playwright>=1.45.0
playwright-stealth>=2.0.0  # stealth patches for Playwright
camoufox>=0.4.0  # Firefox-based stealth browser for Cloudflare/PerimeterX bypass
curl_cffi>=0.7.0  # TLS/JA3 fingerprint impersonation for anti-bot bypass

# Environment and configuration
python-dotenv>=1.0.1  # for environment variable management

# API server for RAG
fastapi>=0.111.0
uvicorn>=0.30.0

# Built-in Python modules (no installation required):
# - os, sys, time, threading, queue, random, uuid
# - csv, glob, re, html, logging
# - argparse, concurrent.futures
# - datetime, collections, typing
# - urllib.parse
# - json, hashlib

# Machine learning model training
lightgbm>=4.0.0  # LightGBM for gradient boosting
optuna>=3.0.0  # Hyperparameter optimization
scikit-learn>=1.3.0  # ML utilities (calibration, metrics, preprocessing)
shap>=0.42.0  # SHAP for model interpretability
matplotlib>=3.5.0  # Plotting for SHAP summaries
joblib>=1.3.0  # Model serialization
pyarrow>=14.0.0  # Parquet export support

# Calibration
venn-abers>=0.4.0  # Venn-ABERS calibration for better-calibrated probabilities

# RAG evaluation (optional — only needed for running eval suite)
# ragas>=0.2.0  # RAGAS metrics (faithfulness, relevancy, context precision)

# Optional but recommended for better performance:
# - lxml (already included above)
# - ujson (for faster JSON processing, included with pymilvus)
# - pandas (for data manipulation, included with pymilvus)
